refactor: Extract useUnityLifecycle hook from UnityView
